Casting the args received.

+++ example-simple/src/ofApp.cpp
@@ -163,7 +163,7 @@ void  ofApp::render(vr::Hmd_Eye nEye)
 //--------------------------------------------------------------
 void ofApp::controllerEvent(ofxOpenVRControllerEventArgs& args)
 {
-	cout << "ofApp::controllerEvent > role: " << args.controllerRole << " - event type: " << args.eventType << " - button type: " << args.buttonType << " - x: " << args.analogInput_xAxis << " - y: " << args.analogInput_yAxis << endl;
+	cout << "ofApp::controllerEvent > role: " << (int)args.controllerRole << " - event type: " << (int)args.eventType << " - button type: " << (int)args.buttonType << " - x: " << args.analogInput_xAxis << " - y: " << args.analogInput_yAxis << endl;
 }
 
 //--------------------------------------------------------------
